@@ -2,21 +2,23 @@
 name = "bpm-ai-experimental"
 version = "0.1.0"
 description = ""
-authors = ["Benjoyo <bennet.krause@gmx.de>"]
+authors = ["Bennet Krause <bennet.krause@holisticon.de>"]
 readme = "README.md"
 
 [tool.poetry.dependencies]
-python = "^3.10"
+python = "^3.12"
 bpm-ai-core = { path = "../bpm-ai-core", develop = true }
-playwright = "^1.39.0"
+playwright = "^1.40.0"
 lxml = "^4.9.3"
 beautifulsoup4 = "^4.12.2"
-pytest = "^7.4.2"
-pytest-dotenv = "^0.5.2"
 
 
 [tool.poetry.group.dev.dependencies]
 setuptools = "^68.2.2"
+
+[tool.poetry.group.test.dependencies]
+pytest = "^7.4.2"
+pytest-dotenv = "^0.5.2"
 
 [build-system]
 requires = ["poetry-core"]
