@@ -3,8 +3,8 @@ from typing import List, Dict, Union, Any
 from PIL.Image import Image
 from openai.types.chat import ChatCompletionMessageParam
 
-from bpm_ai_core.util.image import base64_encode_image
 from bpm_ai_core.llm.common.message import ChatMessage, ToolCallsMessage, ToolResultMessage
+from bpm_ai_core.util.image import base64_encode_image
 
 
 def get_openai_tool_call_dict(message: ToolCallsMessage):
@@ -49,12 +49,13 @@ def message_to_openai_dict(message: ChatMessage) -> ChatCompletionMessageParam:
                     "Elements in ChatMessage.content must be of type str or PIL.Image."
                 )
     else:
-        raise ValueError(
+        content = None
+        print(
             "ChatMessage.content must be of type str or List[Union[str, PIL.Image]] if used for chat completions."
         )
     return {
         "role": message.role,
-        "content": content,
+        **({"content": content} if content else {}),
         **extra_dict,
         **({"name": message.name} if message.name else {})
     }
@@ -79,53 +80,12 @@ def messages_to_openai_dicts(messages: List[ChatMessage]):
     return [message_to_openai_dict(m) for m in messages]
 
 
-def json_schema_from_shorthand(schema: dict) -> dict:
-    def type_or_default(x):
-        # If x is a string, return the string type
-        if isinstance(x, str):
-            return {"type": "string", "description": x}
-
-        # If x is a list, treat it as an array
-        elif isinstance(x, list):
-            # If the list has items, process the first item
-            if x:
-                return {"type": "array", "items": type_or_default(x[0])}
-            else:
-                return {"type": "array", "items": {}}
-
-        # If x is a dictionary
-        elif isinstance(x, dict):
-            # Check for explicitly provided "type" and "properties"
-            if 'type' in x and x['type'] == 'object' and 'properties' in x:
-                properties = {k: type_or_default(v) for k, v in x['properties'].items()}
-                return {"type": "object", "properties": properties, "required": list(properties.keys())}
-
-            # If 'type' is not present or if 'type' is 'object', treat x as an object
-            elif 'type' not in x or x['type'] == 'object':
-                properties = {k: type_or_default(v) for k, v in x.items()}
-                return {"type": "object", "properties": properties, "required": list(properties.keys())}
-            else:
-                return x
-        else:
-            return x
-
-    return {k: type_or_default(v) for k, v in schema.items()}
-
-
-def schema_from_properties(properties: Dict[str, Union[str, dict]]):
-    return {
-        "type": "object",
-        "properties": json_schema_from_shorthand(properties),
-        "required": list(properties.keys()),
-    }
-
-
 def json_schema_to_openai_function(name: str, desc: str, schema: Dict[str, Any]) -> dict:
     return {
         "type": "function",
         "function": {
             "name": name,
             "description": desc,
-            "parameters": schema_from_properties(schema),
+            "parameters": schema
         }
     }
