@@ -1,27 +1,60 @@
 package io.holunda.connector
 
 import com.fasterxml.jackson.module.kotlin.*
-import io.camunda.connector.api.error.BpmnError
 import io.camunda.connector.runtime.core.*
-import py4j.*
-import java.util.Optional
+import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
+import io.ktor.serialization.jackson.*
+import io.ktor.server.application.*
 
-fun main() = GatewayServer(ConnectorRuntime()).start()
+import io.ktor.server.engine.*
+import io.ktor.server.netty.*
+import io.ktor.server.plugins.contentnegotiation.*
+import io.ktor.server.request.*
+import io.ktor.server.response.*
+import io.ktor.server.routing.*
 
-class ConnectorRuntime {
+fun main() {
+    embeddedServer(Netty, port = 9999) { module() }.start(wait = true)
+}
 
-    fun createOutputVariables(context: String, resultExpression: String): Map<String, Any> {
+fun Application.module() {
+    install(ContentNegotiation) {
+        jackson {}
+    }
+
+    routing {
+        post("/createOutputVariables") {
+            val request = call.receive<CreateOutputVariablesRequest>()
+            val result = ConnectorRuntime.createOutputVariables(request.context, request.resultExpression)
+            call.respond(result)
+        }
+
+        post("/examineErrorExpression") {
+            val request = call.receive<ExamineErrorExpressionRequest>()
+            val result = ConnectorRuntime.examineErrorExpression(request.context, request.errorExpression)
+            call.respondNullable(result)
+        }
+    }
+}
+
+data class CreateOutputVariablesRequest(val context: Any, val resultExpression: String)
+data class ExamineErrorExpressionRequest(val context: Any, val errorExpression: String)
+
+
+object ConnectorRuntime {
+
+    fun createOutputVariables(context: Any, resultExpression: String): Map<String, Any> {
         return ConnectorHelper.createOutputVariables(
-            jacksonObjectMapper().readValue(context),
+            context,
             null,
             resultExpression
         )
     }
 
-    fun examineErrorExpression(context: String, errorExpression: String): String {
+    fun examineErrorExpression(context: Any, errorExpression: String): Any? {
         return ConnectorHelper.examineErrorExpression(
-            jacksonObjectMapper().readValue(context),
+            context,
             mapOf("errorExpression" to errorExpression)
-        ).map { it.code }.orElse("None")
+        ).map { it }.orElse(null)
     }
 }
